queue fix 3

+++ database/migrations/2025_01_20_172209_create_chats_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('chats', function (Blueprint $table) {
             $table->id();
-            $table->integer('chat_id')->unique();
+            $table->unsignedBigInteger('chat_id')->unique();
             $table->timestamps();
         });
     }
